- fixed some errors in level loading - improved unit tests, now we can specify many folders for levels

+++ Assets/Scripts/Loading/D2Legacy/DS1Loader.cs
@@ -190,10 +190,11 @@ namespace Diablo2Editor
                 }
             }
 
-            // skip 2 dwords ?
+            
             if ((level.version >= 9) && (level.version <= 13))
             {
-                streamPosition += sizeof(ulong) * 2;
+                // skip 2 ints( 8 bytes)
+                streamPosition += sizeof(int) * 2;
             }
